Feat: Enable support to submit single commits
Each commit in the pull request are processed individually as a
single commit before submitting to Gerrit repository. This option
allows you to preserve git history.

This requires `inputs.SUBMIT_SINGLE_COMMITS` to be set to 'true'
from the caller.

Remove extra inputs. Gerrit server and port and project has be read
from .gitreview file. Composite actions have a limitation on
the number of the inputs that can be passed to 10.

Update doc example with new feat options.

Signed-off-by: Anil Belur <abelur@linuxfoundation.org>