Fix: Reuse Change-Id when PR is reopened
The change-id is added as a comment on the PR and
retrived next time the PR is updated or reopened when
commits are reworked.

This is supported only for squashed commits and does not
work when SUBMIT_SINGLE_COMMITS option is enabled. With
SUBMIT_SINGLE_COMMITS its the developers responsibility to
ensure Change-Id's are reused.

Signed-off-by: Anil Belur <abelur@linuxfoundation.org>