Fix: Re-use change-id takes precedence
If the change has been updated then the change-Id from the PR
takes precedence. Conditional check if the change-Id has
been retrived from the previous fetch step and override
the change-Id provided by the commit message if that exists.

This ensure multiple changes are not created when the change
is updated.

Signed-off-by: Anil Belur <abelur@linuxfoundation.org>